[#14] fix: cloudinary 유틸 파일 생성 및 관련 유틸 함수 이동
